refactor(rules): address review — split markTarget from toggleEffect, drop duration
- Extract target-marking into its own markTarget rule; leave #805's
  toggleEffect foundation untouched (resolves the type-name collision)
- Drop durationDays/worldTime expiry: 'until you mark another' is capacity
  eviction; figurative in-game 'for 1 day' is GM-adjudicated, not tracked
- Generic 'Conditional Bonuses' dialog heading (was 'Quarry Bonus')
- Reword 'marked' condition desc so it no longer claims rule mechanics
- Pack JSONs use markTarget; #utils alias; en.json + inventory updated
