+++ src/utils/conditionalBonuses.test.ts
@@ -181,8 +181,6 @@ describe('conditionalBonuses', () => {
 							actorUuid: 'Actor.t',
 							tokenUuid: null,
 							name: 'Goblin',
-							markedAt: 0,
-							durationDays: null,
 						},
 					],
 				}),
@@ -206,8 +204,6 @@ describe('conditionalBonuses', () => {
 							actorUuid: 'Actor.t',
 							tokenUuid: null,
 							name: 'Goblin',
-							markedAt: 0,
-							durationDays: null,
 						},
 					],
 				}),
